Conflicts:
	data/graphics/images/login_wallpaper.png
	src/being.cpp
	src/beingmanager.cpp
	src/engine.cpp
	src/game.cpp
	src/gui/buysell.cpp
	src/gui/buysell.h
	src/gui/gui.h
	src/gui/npc_text.cpp
	src/gui/npc_text.h
	src/gui/npcintegerdialog.cpp
	src/gui/npclistdialog.cpp
	src/gui/npclistdialog.h
	src/gui/npcstringdialog.cpp
	src/gui/sell.cpp
	src/gui/shop.cpp
	src/gui/table.cpp
	src/net/beinghandler.cpp
	src/net/npchandler.cpp
	src/net/playerhandler.cpp
	src/npc.cpp
	src/npc.h
	src/shopitem.cpp
	src/shopitem.h
	src/utils/stringutils.cpp
	src/utils/stringutils.h
	src/utils/trim.h

+++ src/gui/buy.cpp
@@ -119,7 +119,7 @@ void BuyDialog::reset()
     setMoney(0);
 }
 
-void BuyDialog::addItem(short id, int price)
+void BuyDialog::addItem(int id, int price)
 {
     mShopItems->addItem(id, price);
     mShopItemList->adjustSize();
